Create README.md
This is a code that has not been successful. The purpose of this code is to solve the first principle of wordrop, which is user equilibrium. There are some problems in the code, so the solution does not produce the correct result. 

The object function image drawn by the function ‘fmincon ’shows that this function finds a maximum value. I really don't know where the problem is. I hope that any master will find the problems and give me some advice.